🛠[회원가입_회원생성] :회원생성 컨테이너 생성
